UPDATE ::: Fixing null dereferencing issues
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@38533 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 72d1a938db040ff61dd168d1d04ec4c2e266034d

+++ src/prerna/sablecc2/reactor/frame/r/util/RJavaRemoteRserveTranslator.java
@@ -302,7 +302,9 @@ public class RJavaRemoteRserveTranslator extends RJavaRserveTranslator {
 				}
 			} finally {
 				f.delete();
-				outputF.delete();
+				if (outputF != null) {
+					outputF.delete();
+				}
 			}
 		}
 
@@ -311,7 +313,7 @@ public class RJavaRemoteRserveTranslator extends RJavaRserveTranslator {
 		this.executeEmptyR("gc()");
 
 		if (scriptOutput == null) {
-			//			throw new NullPointerException("Neccesity to trim, scriptOutput cannot be null here.");
+			// throw new NullPointerException("Neccesity to trim, scriptOutput cannot be null here.");
 			return "";
 		}
